requirements.txt
python-telegram-bot==13.15
schedule
python-dotenv
pandas